+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed incomplete data being returned by `IGMarkets::DealingPlatform::AccountMethods#activities` and
   `IGMarkets::DealingPlatform::AccountMethods#transactions`
+- Fixed errors calling `IGMarkets::DealingPlatform#instantiate_models` for models that have deprecated attributes
 
 ### 0.7 — 1 May, 2016